requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
astroid==3.3.8
attrs==24.3.0
black==24.10.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
dill==0.3.9
distro==1.9.0
Flask==3.1.0
Flask-Cors==5.0.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.37.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jsonify==0.5
lint==1.2.1
MarkupSafe==3.0.2
mccabe==0.7.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.1
oauthlib==3.2.2
openai==1.58.1
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
priority==2.0.0
propcache==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
pylint==3.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
Quart==0.20.0
quart-cors==0.8.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
tomlkit==0.13.2
tqdm==4.67.1
twilio==9.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
